feat: implemented question create (#154,#164)

+++ internal/adapter/handler/http/v1/tag.go
@@ -97,7 +97,7 @@ func (h *tagHandler) createMultiple(w http.ResponseWriter, r *http.Request) {
 	for _, rt := range req.Tags {
 		t = append(t, tag.Tag{
 			Name:       rt.Name,
-			LanguageId: rt.LanguageId,
+			LanguageId: uint8(rt.LanguageId),
 		})
 	}
